refactor: rename admin-api.ts to agents-api.ts
Now that the /admin/api/ path prefix is gone (UNI-1.1), the file name
admin-api.ts is a misnomer. Rename to agents-api.ts to match the unified
/agents/* route surface it serves.

Files renamed:
  admin/src/admin-api.ts → admin/src/agents-api.ts
  admin/src/admin-api.smoke.test.ts → admin/src/agents-api.smoke.test.ts
  admin/src/admin-api.integration.test.ts → admin/src/agents-api.integration.test.ts

Updated all import references in server.ts, main.ts, api.smoke.test.ts,
and docs/agent.md.
