+++ admin/src/api.smoke.test.ts
@@ -10,7 +10,7 @@
 import { describe, expect, test } from "bun:test";
 import { Hono } from "hono";
 import { sign } from "hono/jwt";
-import { createAdminApp, parseAdminApiKeys } from "./admin-api.ts";
+import { createAdminApp, parseAdminApiKeys } from "./agents-api.ts";
 import type { AgentCronJob } from "./agent-cron-jobs.ts";
 import type { AgentEnvBundle } from "./agent-envs.ts";
 import { createAgentRuntimeApp } from "./api.ts";
